requirements.txt
flask==3.0.3
flask-cors==4.0.1
anthropic==0.49.0
gunicorn==23.0.0
dotenv==0.21.0
google-generativeai
gunicorn
google-genai